+++ packages/apollo-angular/src/api.ts
@@ -2,3 +2,4 @@ export {Apollo} from './Apollo';
 export {QueryRef} from './QueryRef';
 export {SelectPipe} from './SelectPipe';
 export {ApolloModule} from './ApolloModule';
+export {ApolloOptionsToken, ApolloNamedOptionsToken} from './tokens';